refactor: trim comments and tests, fix test filename convention
Reduce 13-test file to 3 focused tests (one per distinct bug: datetime.time
kwargs fix, timedelta dispatch fix, PyComplex_Check). Remove redundant edge-case
tests and sanity checks for unrelated pre-existing behaviour. Rename test file
from hyphens to underscores per validate-test-names convention.
